feature: [#37] Util 클래스 수정
- File 유틸 클래스 생성
- Image 유틸 클래스 패키지 이동
- S3 유틸 클래스 수정
