+++ src/main/java/com/tinyhuman/tinyhumanapi/integration/aws/S3Util.java
@@ -1,28 +1,6 @@
 package com.tinyhuman.tinyhumanapi.integration.aws;
 
 public class S3Util {
-
-    public static String addUserIdToImagePath(String path, Long userId, String fileName) {
-        StringBuilder s3FullPathBuilder = new StringBuilder();
-        String id = String.valueOf(userId);
-        return s3FullPathBuilder.append(path.replace("userId", id)).append(fileName).toString();
-    }
-
-    public static String addBabyIdToImagePath(String path, Long babyId, String fileName) {
-        StringBuilder s3FullPathBuilder = new StringBuilder();
-        String id = String.valueOf(babyId);
-        return s3FullPathBuilder
-                .append(path.replace("babyId", id))
-                .append(fileName).toString();
-    }
-
-    public static String addBabyIdAndAlbumIdToImagePath(String path, Long babyId, Long albumId, String fileName) {
-        StringBuilder s3FullPathBuilder = new StringBuilder();
-        return s3FullPathBuilder
-                .append(path.replace("babyId", String.valueOf(babyId)).replace("albumId", String.valueOf(albumId)))
-                .append(fileName).toString();
-    }
-
     public static String getAccessUrl(String bucketName, String s3FullPath) {
         StringBuffer accessUrl = new StringBuffer();
         return accessUrl.append("https://")
